+++ ecmascript-testcases/test-dev-exec-timeout-tailcall.js
@@ -1,0 +1,24 @@
+/*
+ *  Execution timeout: tailcall function
+ */
+
+/*---
+{
+    "skip": true
+}
+---*/
+
+// See: GH-214
+function f() {
+    return f();
+}
+
+function test() {
+    f();
+}
+
+try {
+    test();
+} catch (e) {
+    print(e.name);
+}